further steps on allowing UI to create scapes and have other sections of the UI respond automatically to those scapes
In the compose dialog you can now create new scapes.  And within those scapes you can set scape entires.

Additionally the if you create scape with a droplet-address->boolean relationship then the UI will add button next to a droplet so it can be categorized into that scape.

Also added a scape to keep track of the user created scapes.