added channel-type scape to streamscapes and changed the droplet-channel scape to relate droplet addresses with channel addresses rather than channel names.
This channel-type is populated by the new->scape signal and is now used by the UI to determine scape types instead of the silly hack with the scape names.  This also allows some refactoring of channel-type code into ss-utils and means that it will be possible in the future to change a channel's name simply by updating the channel scape instead of having to go and also update all the entries in the droplet-channel scape.