many compository and improvements
grooves are specify their grammar just once, and separately which channels can can carry them, as well as how to match for that groove on the grammar, and also what elements of the content (or the groove match) to display in the preview.

Updated the UI to use this new info and display previews based on the groove definition.